refactor(live-sessions): extract active-session-game-scene hooks
Split active-session-game-scene.tsx effects into dedicated hooks:
- use-tournament-detail.ts: composite query hook (tournament + chip purchases
  + blind levels + currencies) formerly defined inline
- use-ring-game-scene-actions.ts: ring game update + dialog state + session
  invalidation for the cash game branch
- use-tournament-scene-actions.ts: tournament saveWithLevels + dialog state
  + related invalidations for the tournament branch

Pure formatters (variantLabel, formatBlindParts, formatAnteSuffix,
VARIANT_LABELS) move to game-scene-formatters utils.

Both scene-action hooks use invalidateTargets from optimistic-update.
Presentational sub-components (GameSceneShell, RingGameDetailsCard,
TournamentInfoCard, ChipPurchasesCard, StructureCard,
TournamentStructureTable) stay co-located.
